--- test/e2e/app-dir/import-trace-dynamic-eval/app/route.ts
@@ -1,8 +1,0 @@
-import { foo } from '../lib/foo'
-
-export function GET() {
-  foo()
-  return Response.json('Hello World!')
-}
-
-export const runtime = 'edge'
