--- test/e2e/app-dir/import-trace-dynamic-eval/middleware.ts
@@ -1,5 +1,0 @@
-import { foo } from './lib/foo'
-
-export function middleware() {
-  foo()
-}
