+++ test/production/edge-dynamic-code-eval/lib/foo.js
