+++ test/production/edge-dynamic-code-eval/pages/index.js
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <p>hello world</p>
+}
